Relocate App.js and App.css to App folder
