+++ src/App/App.js
@@ -3,7 +3,7 @@ import './App.css';
 const App = () => {
   return (
     <div>
-
+      <h1> Oscar Cortez Barrientos</h1>
     </div>
   )
 }
